VMC 2.x - Remove unnecessary usage of getProcessOutput
With Python 2.6 and unpatched glib2 it triggers 100% CPU. At present
only Ubuntu appears to be applying a patch to glib2 but Fedora, Moblin
and OpenSUSE are known to suffer this issue. This doesn't entirely
fix the problem since getProcessOutput is still used to start
external commands like Internet browser and mail tool etc. Switching
to use os.spawn or subprocess.Popen doesn't help as the issue appears
to be the reactor's use of a SIGCHLD handler.